+++ src/components/mobile/overlay/mobile-overlay-panel.component.tsx
@@ -1,11 +1,13 @@
-import { CSSProperties, useEffect, useRef, useState } from 'react';
+import React, { useEffect, useRef, useState } from 'react';
 import ReactDOM from 'react-dom';
 import { bindDocumentClickListener, unbindDocumentClickListener } from '../../common-event-handlers.ts';
 import { getAllElementStyleValues } from '../../common-functions.ts';
-import BeeSoftTransition from '../../common/beesoft-transition/beesoft-transition.component.tsx';
 import { Button } from '../../navigation/buttons/button/button.component.tsx';
 import { getTargetElement } from '../../overlay/overlay-functions.ts';
 import { MobileOverlayPanelProps } from './mobile-overlay-panel.props.ts';
+import { useEvent } from "@beesoft/common";
+import { TargetAndTransition } from "motion";
+import { motion } from "motion/react";
 
 const MobileOverlayPanel = ({
   visible,
@@ -13,7 +15,7 @@ const MobileOverlayPanel = ({
   appendTo = document.body,
   unmountWhenHidden = false,
   transitionDuration = 400,
-  showTransitionOptions = 'cubic-bezier(0, 0, 0.2, 1)',
+  showTransitionOptions = 'easeInOut',
   hideTransitionOptions = 'linear',
   approveText = 'DONE',
   shown,
@@ -24,12 +26,16 @@ const MobileOverlayPanel = ({
   const [zIndex, setZIndex] = useState(-1);
   const [underlayZIndex, setUnderlayZIndex] = useState(-1);
   const [visibility, setVisibility] = useState(visible);
+  const [animationComplete, setAnimationComplete] = useState(true);
 
   const underlayDisplayZIndex = useRef(100);
   const displayZIndex = useRef(101);
-  const finalTarget = useRef<HTMLElement>();
-  const panelRef = useRef<HTMLElement>();
-  const listenerRef = useRef<(event: MouseEvent) => void>();
+  const finalTarget = useRef<HTMLElement>(undefined);
+  const panelRef = useRef<HTMLElement>(undefined);
+  const listenerRef = useRef<(event: MouseEvent) => void>(undefined);
+  const hasShown = useRef(false);
+
+  const durationInSeconds = transitionDuration / 1000;
 
   useEffect(() => {
     if (target) {
@@ -49,6 +55,7 @@ const MobileOverlayPanel = ({
       displayZIndex.current = Math.max(...parentZIndex) + 2;
     }
 
+    setAnimationComplete(false);
     setVisibility(visible);
   }, [visible, target]);
 
@@ -59,6 +66,7 @@ const MobileOverlayPanel = ({
   const onEntering = () => {
     setUnderlayZIndex(underlayDisplayZIndex.current);
     setZIndex(displayZIndex.current);
+    hasShown.current = true;
   };
 
   const onEntered = () => {
@@ -75,6 +83,7 @@ const MobileOverlayPanel = ({
       if (clickedWithin) {
         isClickedWithin?.();
       } else {
+        setAnimationComplete(false);
         setVisibility(false);
       }
     };
@@ -94,52 +103,78 @@ const MobileOverlayPanel = ({
     hidden?.();
   };
 
-  const maskOverrideStyles: Record<string, CSSProperties> = {
-    entering: { opacity: 0.5 },
-    entered: { opacity: 0.5 },
+  const onAnimationStart = useEvent(() => {
+    if (visibility) {
+      onEntering();
+    } else {
+      onExit();
+    }
+  });
+
+  const onAnimationEnd = useEvent(() => {
+    if (visibility) {
+      onEntered();
+    } else {
+      onExited();
+    }
+
+    setAnimationComplete(true);
+  });
+
+  const underlayVisibleState: TargetAndTransition = {
+    opacity: 0.5,
   };
 
-  return ReactDOM.createPortal(
-    <BeeSoftTransition
-      start={visibility}
-      timeout={transitionDuration}
-      showTransitionOptions={showTransitionOptions}
-      hideTransitionOptions={hideTransitionOptions}
-      onEntering={onEntering}
-      onEntered={onEntered}
-      onExit={onExit}
-      onExited={onExited}
-      unmountOnExit={unmountWhenHidden}
-    >
-      {({ state, defaultStyle, transitionStyles }) => (
-        <div ref={(element) => element && onPanelCreated(element)}>
-          <div
-            className="bsc:fixed bsc:left-0 bsc:top-0 bsc:h-full bsc:w-full bsc:bg-mono-dark-1"
-            style={{
-              zIndex: underlayZIndex,
-              ...transitionStyles[state],
-              ...maskOverrideStyles[state],
-            }}
-          />
-          <div
-            className="dark:bsck-border-mono-light-1 bsc:fixed bsc:bottom-0 bsc:left-0 bsc:w-full bsc:bg-white bsc:dark:border-t bsc:dark:border-solid bsc:dark:bg-mono-dark-1 bsc:dark:text-mono-light-1"
-            style={{ zIndex, ...defaultStyle, ...transitionStyles[state] }}
+  const underlayHiddenState: TargetAndTransition = {
+    opacity: 0,
+  };
+
+  const panelVisibleState: TargetAndTransition = {
+    opacity: 1,
+    pointerEvents: 'auto',
+  };
+
+  const panelHiddenState: TargetAndTransition = {
+    opacity: 0,
+    pointerEvents: 'none',
+  };
+
+  return hasShown.current && !visibility && unmountWhenHidden === true && animationComplete
+    ? null
+    : ReactDOM.createPortal(
+    <div>
+      <motion.div
+        className="bsc:fixed bsc:left-0 bsc:top-0 bsc:h-full bsc:w-full bsc:bg-mono-dark-1"
+        style={{ zIndex: underlayZIndex }}
+        initial={false}
+        animate={visibility ? underlayVisibleState : underlayHiddenState}
+      />
+      <motion.div
+        ref={(element) => { if (element) onPanelCreated(element) }}
+        className="dark:bsc-border-mono-light-1 bsc:fixed bsc:bottom-0 bsc:left-0 bsc:w-full bsc:bg-white bsc:dark:border-t bsc:dark:border-solid bsc:dark:bg-mono-dark-1 bsc:dark:text-mono-light-1"
+        style={{ zIndex }}
+        initial={false}
+        animate={visibility ? panelVisibleState : panelHiddenState}
+        transition={{
+          duration: durationInSeconds,
+          ease: visibility ? showTransitionOptions : hideTransitionOptions
+        }}
+        onAnimationStart={onAnimationStart}
+        onAnimationComplete={onAnimationEnd}
+      >
+        <div className="bsc:w-full">{children}</div>
+        <>
+          <Button
+            buttonType="primary"
+            fullWidth={true}
+            onClick={() => setVisibility(false)}
+            className="bsc:text-xs bsc:font-bold"
           >
-            <div className="bsc:w-full">{children}</div>
-            <>
-              <Button
-                buttonType="primary"
-                fullWidth={true}
-                onClick={() => setVisibility(false)}
-                className="bsc:text-xs bsc:font-bold"
-              >
-                {approveText}
-              </Button>
-            </>
-          </div>
-        </div>
-      )}
-    </BeeSoftTransition>,
+            {approveText}
+          </Button>
+        </>
+      </motion.div>
+    </div>,
     appendTo
   );
 };